+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 
 ### Added
 
+* Support catching panics, and raising JS Exceptions for them, when building
+  with panic=unwind on nightly, with the new `catch-unwind` feature.
+  [#4790](https://github.com/wasm-bindgen/wasm-bindgen/pull/4790)
+
 * Added `private` attribute on exported types to allow generating
   exports and structs as implicit internal exported types for function
   arguments and returns, without exporting them on the public interface.